Correct the assignment of data tag to bookcards
The data tag is intended to store the index in the myLibrary array of
the book displayed by the book card. This data tag was off by one,
which caused desyncs between the displayed library and the library
array. Now the data tags correctly list the indeces of the books